pyproject.toml
[project]
name = "browser-butler"
version = "0.1.0"
description = "AI-powered browser automation with scheduling and Telegram notifications"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "apscheduler>=3.10.0",
    "browser-use>=0.1.0",
    "playwright>=1.40.0",
    "aiogram>=3.3.0",
    "pydantic-settings>=2.1.0",
    "openai>=1.10.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
